refactor: API 파일 도메인별 통합 (#29 - Step 1)
13개의 개별 API 파일을 7개의 도메인 파일로 통합

- articles.js: schoolArticles + clubArticles + getHotEventList 통합
- calendar.js: getMonthlyAll 이동
- bookmarks.js: getSchoolBookmarks + postBookmark + delete* 통합
- user.js: patchUserMajor + deleteAccount 통합
- auth.js: postGoogleLogin 이동
- vendors.js: getVendors 이동
- notifications.js: 기존 유지
