+++ CLAUDE.md
@@ -7,7 +7,7 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 ## Claude 행동 지침
 
 1. **CLAUDE.md 자동 업데이트**: 큰 기능이 구현되거나 프로젝트 구조/컨벤션이 변경되면 이 파일도 함께 수정한다.
-2. **학습 설명 포함**: 코드 구현 시 사용자의 웹 프론트엔드 학습을 위해 구현한 코드에 대한 이론적 설명을 함께 제공한다.
+2. **학습 설명 포함**: 코드 구현 시 사용자의 학습을 위해 구현한 코드에 대한 구체적인 설명을 함께 제공한다.
 3. **단계별 구현**: 코드 구현 시 과정을 단계별로 나누고 한 단계가 끝나면 사용자에게 다음 단계 진행을 허가받은 후 다음 단계로 넘어간다.
 
 ---
