+++ src/components/main/adaptive/feature/MYP/BookmarkSection.jsx
@@ -2,9 +2,7 @@ import React from "react";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { useNavigate } from "react-router-dom";
 import BookmarkItem from "@/components/main/adaptive/feature/MYP/BookmarkItem";
-import { deleteSchoolBookmarksAll } from "@/api/deleteSchoolBookmarksAll";
-import { deleteSchoolBookmark } from "@/api/deleteSchoolBookmark";
-import { getSchoolBookmarks } from "@/api/getSchoolBookmarks";
+import { deleteSchoolBookmarksAll, deleteSchoolBookmark, getSchoolBookmarks } from "@/api/bookmarks";
 import { useDeviceStore } from "@/stores/deviceStore";
 import useAuthStore from "@/stores/useAuthStore";
