+++ src/pages/main/LGN/LGNPage.jsx
@@ -2,7 +2,7 @@ import { useNavigate, useLocation } from "react-router-dom";
 import useAuthStore from "@/stores/useAuthStore";
 import { useDeviceStore } from "@/stores/deviceStore";
 import { GoogleLogin } from "@react-oauth/google";
-import { postGoogleLogin } from "@/api/postGoogleLogin";
+import { postGoogleLogin } from "@/api/auth";
 const LGNPage = () => {
   const login = useAuthStore((state) => state.login);
   const isMobile = useDeviceStore((state) => state.isMobile);
